Add files via upload
A scientific calculator is a type of electronic calculator, designed to calculate problems in science, engineering, and mathematics .Scientific calculators are used widely in situations that require quick access to certain mathematical functions. They are also used for calculations of very large or very small numbers, as in some aspects of astronomy, physics, and chemistry,and are widely used in both education and professional purpose.